Update README.md
This project is a simple and detailed guide for everyone who wants to learn how to use a computer or laptop. Here you will find easy lessons, useful tips, and practical exercises for beginners and those who want to improve their skills. The main goal is to make working with computers clear and accessible for everyone!